chore: cleanup RelativeFlag
